Delete MapUtils.
